refactor: rename ish-filter-collapsible component
