+++ src/app/shared/components/filter/filter-navigation-sidebar/filter-navigation-sidebar.component.spec.ts
@@ -5,7 +5,7 @@ import { FilterNavigation } from 'ish-core/models/filter-navigation/filter-navig
 import { Filter } from 'ish-core/models/filter/filter.model';
 import { findAllCustomElements } from 'ish-core/utils/dev/html-query-utils';
 import { FilterCheckboxComponent } from 'ish-shared/components/filter/filter-checkbox/filter-checkbox.component';
-import { FilterCollapsableComponent } from 'ish-shared/components/filter/filter-collapsable/filter-collapsable.component';
+import { FilterCollapsibleComponent } from 'ish-shared/components/filter/filter-collapsible/filter-collapsible.component';
 import { FilterDropdownComponent } from 'ish-shared/components/filter/filter-dropdown/filter-dropdown.component';
 import { FilterSwatchImagesComponent } from 'ish-shared/components/filter/filter-swatch-images/filter-swatch-images.component';
 import { FilterTextComponent } from 'ish-shared/components/filter/filter-text/filter-text.component';
@@ -22,7 +22,7 @@ describe('Filter Navigation Sidebar Component', () => {
       declarations: [
         FilterNavigationSidebarComponent,
         MockComponent(FilterCheckboxComponent),
-        MockComponent(FilterCollapsableComponent),
+        MockComponent(FilterCollapsibleComponent),
         MockComponent(FilterDropdownComponent),
         MockComponent(FilterSwatchImagesComponent),
         MockComponent(FilterTextComponent),
@@ -53,34 +53,34 @@ describe('Filter Navigation Sidebar Component', () => {
     } as FilterNavigation;
 
     fixture.detectChanges();
-    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsable', 'ish-filter-dropdown']);
+    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsible', 'ish-filter-dropdown']);
   });
 
   it('should display filter-text if facet with displayType text_clear is present', () => {
     component.filterNavigation = { filter: [{ displayType: 'text_clear' } as Filter] } as FilterNavigation;
 
     fixture.detectChanges();
-    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsable', 'ish-filter-text']);
+    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsible', 'ish-filter-text']);
   });
 
   it('should display filter-swatch-images if facet with displayType swatch is present', () => {
     component.filterNavigation = { filter: [{ displayType: 'swatch' } as Filter] } as FilterNavigation;
 
     fixture.detectChanges();
-    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsable', 'ish-filter-swatch-images']);
+    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsible', 'ish-filter-swatch-images']);
   });
 
   it('should display filter-text if facet has no displayType set', () => {
     component.filterNavigation = { filter: [{} as Filter] } as FilterNavigation;
 
     fixture.detectChanges();
-    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsable', 'ish-filter-text']);
+    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsible', 'ish-filter-text']);
   });
 
   it('should display filter-text if facet has a typo in the displayType', () => {
     component.filterNavigation = { filter: [{ displayType: 'typo' } as Filter] } as FilterNavigation;
 
     fixture.detectChanges();
-    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsable', 'ish-filter-text']);
+    expect(findAllCustomElements(element)).toEqual(['ish-filter-collapsible', 'ish-filter-text']);
   });
 });
